Allow user to disable private channels
By default, we will still get private channels. This requires the extra
`groups:` scopes.

Previously, even if we didn't want to sync private groups, the calls to
the Slack API would fail as the calls we made defined the types as
`"public_channel,private_channel"`.

There is now an optional config, `private_channels`, that allows the
user to disable private channels in the config file.

Also added documentation for the config file.